# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Web Profil RW 01 Tunjungsekar",
    "version": "1.0.0",
    "description": "Dokumentasi REST API backend untuk Web Profil RW 01 Kelurahan Tunjungsekar Kota Malang"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/gallery": {
      "get": {
        "summary": "Daftar Foto Galeri (Public)",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "eventId", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "List foto galeri" } }
      }
    },
...
    "/api/v1/admin/gallery/{id}": {
      "delete": {
        "summary": "Hapus Foto Galeri (Admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Foto berhasil dihapus" } }
      }
    },
...
  }
}